--- beacon_chain/transition/src/delegation/mod.rs
@@ -1,6 +1,0 @@
-use super::honey_badger_split;
-use super::types;
-use super::TransitionError;
-use super::shuffling::shuffle;
-
-pub mod validator;
