+++ beacon_chain/validator_induction/src/lib.rs
@@ -10,3 +10,7 @@ pub use inductor::{
     ValidatorInductor,
     ValidatorInductionError,
 };
+
+pub use registration::{
+    ValidatorRegistration,
+};
