+++ beacon_chain/validator_shuffling/src/lib.rs
@@ -1,0 +1,8 @@
+extern crate honey_badger_split;
+extern crate vec_shuffle;
+extern crate types;
+
+mod active_validator_indices;
+mod shuffle;
+
+pub use shuffle::shard_and_committees_for_cycle;
